fix(module:code-editor): only emit update if value changed
Before notifying Angular (through the ControlValueAccessor mechanism) of a change to
the value we make sure that the value has actually changed. This is necessary to
avoid feedback loops as writeValue (model -> view) would otherwise cause this to be
fired (view -> model) which can cause infinite loop issues on the usage side or
broken validation detection.

fixes #5869