fix(module:autocomplete): remove NgZone dependency
This commit eliminates the `NgZone` dependency from the autocomplete component.
Since zone.js is becoming optional, `onStable` won't emit any value when `provideZonelessChangeDetection`
is used in the application config. Instead, we now use the alternative approach provided by `afterNextRender`.
Most of the code that was using `onStable` should be replaced with `afterNextRender`.