refactor(module:core): expose `fromEventOutsideAngular` utility
In this commit, we introduce a helper function, `fromEventOutsideAngular`, as a replacement for
`ngZone.runOutsideAngular + fromEvent`. Using this helper function simplifies things, as there is
no longer a need to inject the `NgZone` service everywhere.

Note that there are no functional changes—only formatting changes, as indentation has been
shifted to the left.